Conflicts:
	.gitignore

+++ src/LibraryBundle/Controller/DefaultController.php
@@ -155,15 +155,24 @@ class DefaultController extends Controller
         return $this->render('LibraryBundle:Default:addCategory.html.twig', array('form'=>$form->createView()));
     }
 
-    public function removeAuthorAction($slug)
+    public function removeAuthorAction(Request $request, $slug)
     {
         $em = $this->getDoctrine()->getManager();
         $repo = $em->getRepository("LibraryBundle:Author");
         $author = $repo->findOneBySlug($slug);
-        $em->remove($author);
-        $em->flush();
 
-        return $this->redirectToRoute('library_homepage');
+        if (null === $author) {
+            throw new NotFoundHttpException("Cet auteur n'existe pas");
+        }
+
+        $form = $this->createFormBuilder()->getForm();
+        if ($form->handleRequest($request)->isValid()) {
+            $em->remove($author);
+            $em->flush();
+            return $this->redirectToRoute('library_homepage');
+        }
+
+        return $this->render('LibraryBundle:Default:removeAuthor.html.twig', array('slug' => $slug, 'form' => $form->createView()));
     }
 
     public function removeBookAction(Request $request, $slug)
